# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Capacitarr API
  version: 1.0.0
  description: |
    Capacitarr is a capacity management tool for media servers. It monitors disk
    usage across Sonarr, Radarr, Lidarr, Readarr, Plex, Jellyfin, Emby, Tautulli,
    and Overseerr integrations, scores media items for potential cleanup based on
    configurable weights and rules, and can automatically (or via approval) free
    disk space when thresholds are breached.

    ## Authentication

    Most endpoints require authentication via one of:
    - **JWT cookie** (`jwt`) — set automatically on login
    - **Bearer token** — `Authorization: Bearer <token>` header
    - **API key** — `X-Api-Key: <key>` header

    The `/health`, `/version`, and `/auth/login` endpoints are public.
  license:
    name: PolyForm Noncommercial 1.0.0
    url: https://polyformproject.org/licenses/noncommercial/1.0.0/
servers:
  - url: http://localhost:2187/api/v1
    description: Local development server
security:
  - cookieAuth: []
  - bearerAuth: []
  - apiKey: []
...
paths:
...
  /audit-log:
    get:
      operationId: listAuditLog
      tags: [Audit Log]
      summary: List audit log entries
      description: |
        Returns paginated audit log entries (permanent history of deletions and dry-runs).
        The default sort is by `created_at` descending (newest first).
      parameters:
        - name: limit
          in: query
          description: Maximum number of records to return (max 1000)
          schema:
            type: integer
            default: 50
            maximum: 1000
        - name: offset
          in: query
          description: Number of records to skip for pagination
          schema:
            type: integer
            default: 0
        - name: search
          in: query
          description: Filter by media name (substring match)
          schema:
            type: string
        - name: action
          in: query
          description: Filter by action type
          schema:
            type: string
            enum: [deleted, dry_run, dry_delete]
        - name: sort_by
          in: query
          description: Column to sort by
          schema:
            type: string
            enum: [created_at, media_name, size_bytes, action]
            default: created_at
        - name: sort_dir
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated audit log response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /audit-log/recent:
    get:
      operationId: getAuditLogRecent
      tags: [Audit Log]
      summary: Get recent audit log entries
      description: Returns the most recent audit log entries for the dashboard mini-feed.
      parameters:
        - name: limit
          in: query
          description: Maximum number of records to return
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /audit-log/grouped:
    get:
      operationId: getAuditLogGrouped
      tags: [Audit Log]
      summary: Get grouped audit log entries
      description: |
        Returns audit log entries grouped by TV show title for season/episode entries.
        Non-TV entries are returned as standalone items. Useful for a tree-style
        audit view.
      parameters:
        - name: limit
          in: query
          description: Maximum number of raw records to fetch before grouping (max 2000)
          schema:
            type: integer
            default: 200
            maximum: 2000
      responses:
        "200":
          description: Grouped audit log results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupedAuditResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ===========================================================================
  # Approval Queue (active items in the approval workflow)
  # ===========================================================================
  /approval-queue:
    get:
      operationId: listApprovalQueue
      tags: [Approval Queue]
      summary: List approval queue items
      description: |
        Returns items in the approval queue. Optionally filter by status.
      parameters:
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Approval queue items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApprovalQueueItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /approval-queue/{id}/approve:
    post:
      operationId: approveQueueItem
      tags: [Approval Queue]
      summary: Approve a queued item for deletion
      description: |
        Approves a pending approval queue item, moving it to "approved"
        status and queuing the associated media item for background deletion.

        Requires that deletions are enabled in settings. If deletions are
        disabled, the request is rejected with a `409 Conflict` response.
      parameters:
        - name: id
          in: path
          required: true
          description: Approval queue item ID
          schema:
            type: integer
      responses:
        "200":
          description: Item approved and queued for deletion
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "approved"
        "400":
          description: Bad request — item is not in pending status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict — deletions are disabled in settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /approval-queue/{id}/reject:
    post:
      operationId: rejectQueueItem
      tags: [Approval Queue]
      summary: Reject (snooze) a queued item
      description: |
        Rejects a pending approval queue item, setting its status to "rejected"
        and applying a snooze period. The item will not appear in the queue
        until the snooze expires or it is manually unsnoozed.
      parameters:
        - name: id
          in: path
          required: true
          description: Approval queue item ID
          schema:
            type: integer
      responses:
        "200":
          description: Item rejected and snoozed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "rejected"
        "400":
          description: Bad request — item is not in pending status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /approval-queue/{id}/unsnooze:
    post:
      operationId: unsnoozeQueueItem
      tags: [Approval Queue]
      summary: Unsnooze a rejected item
      description: |
        Returns a rejected (snoozed) approval queue item to pending status,
        clearing the snooze period. The item will reappear in the queue.
      parameters:
        - name: id
          in: path
          required: true
          description: Approval queue item ID
          schema:
            type: integer
      responses:
        "200":
          description: Item unsnoozed and returned to pending
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "unsnoozed"
        "400":
          description: Bad request — item is not in rejected status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ===========================================================================
  # SSE (Server-Sent Events)
  # ===========================================================================
...
components:
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API key generated via `POST /auth/apikey`
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login response
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt
      description: HttpOnly JWT cookie set on login

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
...
  responses:
...
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
      required: [error]
      example:
        error: "Invalid request body"
...
    AuditLogEntry:
      type: object
      description: A permanent record of a deletion or dry-run action (append-only)
      properties:
        id:
          type: integer
          example: 1
        mediaName:
          type: string
          example: "Breaking Bad - Season 1"
        mediaType:
          type: string
          example: season
        reason:
          type: string
          description: Human-readable score summary
          example: "Score: 0.85 (WatchHistory: 1.0, Size: 0.5)"
        scoreDetails:
          type: string
          description: JSON-encoded array of ScoreFactor objects
          example: '[{"name":"Watch History","rawScore":1.0,"weight":10,"contribution":0.4,"type":"weight"}]'
        action:
          type: string
          enum: [deleted, dry_run, dry_delete]
          example: deleted
        sizeBytes:
          type: integer
          format: int64
          example: 5368709120
        integrationId:
          type: integer
          nullable: true
          description: FK to integration (nullable — preserved when integration is deleted)
          example: 1
        createdAt:
          type: string
          format: date-time
      required: [id, mediaName, mediaType, reason, action, sizeBytes, createdAt]
    AuditLogResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditLogEntry"
        total:
          type: integer
          description: Total number of records matching the filter
          example: 150
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0
      required: [data, total, limit, offset]
    ApprovalQueueItem:
      type: object
      description: An item in the approval queue (state machine — pending/approved/rejected)
      properties:
        id:
          type: integer
          example: 1
        mediaName:
          type: string
          example: "Breaking Bad - Season 1"
        mediaType:
          type: string
          enum: [movie, show, season, episode, artist, album, book]
          example: season
        reason:
          type: string
          description: Human-readable score summary
          example: "Score: 0.85 (WatchHistory: 1.0, Size: 0.5)"
        scoreDetails:
          type: string
          description: JSON-encoded array of ScoreFactor objects
        sizeBytes:
          type: integer
          format: int64
          example: 5368709120
        integrationId:
          type: integer
          description: FK to integration (required)
          example: 1
        externalId:
          type: string
          description: External ID in the source integration
          example: "12345"
        status:
          type: string
          enum: [pending, approved, rejected]
          example: pending
        snoozedUntil:
          type: string
          format: date-time
          nullable: true
          description: When snooze expires (rejected items only)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, mediaName, mediaType, reason, sizeBytes, integrationId, status, createdAt, updatedAt]
...
    GroupedAuditResult:
      type: object
      properties:
        type:
          type: string
          enum: [group, single]
          description: Whether this is a grouped TV entry or a standalone item
        group:
          $ref: "#/components/schemas/AuditGroup"
        entry:
          $ref: "#/components/schemas/AuditLogEntry"
    AuditGroup:
      type: object
      description: A group of audit log entries belonging to the same TV show
      properties:
        showTitle:
          type: string
          example: "Breaking Bad"
        children:
          type: array
          items:
            $ref: "#/components/schemas/AuditLogEntry"
        totalSize:
          type: integer
          format: int64
          description: Combined size of all children
        action:
          type: string
          description: Action of the first child (representative)
        createdAt:
          type: string
          format: date-time
          description: Timestamp of the first child (representative)
      required: [showTitle, children, totalSize, action, createdAt]

    # ---- Data Reset ----
...